fix: wrong package names after moving files
