+++ internal/runmode/runmode.go
@@ -1,4 +1,4 @@
-package main
+package runmode
 
 import (
 	"flag"
